Widget and tooltip instance
WidgetDiv and tooltip are now managed as instances (not one instance for
all workspaces)

So for one workspace, we have one widgetDiv and one tooltipDiv.
Those divs are placed next to svg html tag (for better integration in
websites)

This fix has been realized for better integration.
Then to follow the multiple workspaces behavior, the widget and tooltip
are now associated to the workspace (not as global objects)