Add a dropdown field with a search box.
The addition of this field will alleviate finding of items in dropdown lists for larger programs. One use case is for programs with many variables. Changing the variables_get block to use this field would allow users to filter the list of variables, to find just the variable they need.