Modify the colour_rgb function to match other languages
There are two tests fail before modifying the color_rgb function
because it behaves differently in Dart and in other languages. In Dart,
this function takes parameters ranging from 0 to 1.0 while in other
languages such as Lua the counterpart function takes parameters ranging
from 0 to 100. Now I have modified it to let it behave the same as
other languages.