Store Block hue so it can be extracted reliably
Added HSV hue as a private field to Block so that if
the colour of a block is set to a certain hue, that
same hue can be retrieved. Fixes #1247.

Added a get method and modified the set method.

Also added check to set method to only allow hues between 0 and 360.