Removed incorrect escape sequence
Hello. This line seems to cause any percents in strings to be turned into a literal \%.
This is because a Python percent cannot be escaped using a backslash. 

In a normal Python string I think that a % is meaningless and must not be escaped.

When a percent is used in the context of formatting it should be escaped as %%.