Fix #1960 by collecting block info before icons are destroyed (#1970)
The icons in rendered Blockly currently own the comment text. When
a block was deleted the icons were being disposed of before the block
info was recorded in the event, which meant the comment was lost. This
adds some additional logic to block_svg to make sure the event is
captured before the icons are removed.