Remove goog.ui.ColorPicker
Closure spells “Colour” wrong, so we shouldn’t be using this library.

Adds the ability to set the tooltip titles for each colour.

Note that Closure’s colour picker in Blockly used to have keyboard controls (as of > 3 years ago); but no longer seems to.  This replacement does not currently have keyboard controls either.