Moved highlighting blocks to blockly
Highlighting and centering blocks was done in fable functions.
More specifically, the domain layer, top menu, search.

This is no bueno because we just make the fable functions script
do all of the highlighting and all of the special cases.

Instead, each search handler should be responsible for storing its
results and highlighting a specified single result out of them.