Search - block replacement not working fix
This commit is related to the search logic.
Imagine the following example:

The user has an `if (true)` block in their project.
The `true` boolean is a shadow block.
The user drags a real boolean block from the toolbox and places it
in the place of the shadow block.

What happens underneath should be:
- the new block gets initialized (search trie including)
- the old shadow block gets removed (search trie including)

What happened was:
- the old shadow block removes the new block's entry in the search trie.

This was due to me using `.pop` instead of `.splice` when deregistering
blocks.