Reset the trashcan history on Blockly init finish
This is a bug fix for the trashcan.
Sometimes, Blockly programatically deletes blocks in the workspace
while loading (initializing).

This leads to the trashcan sometimes having the entire project in the
workspace, but also in the trashcan's history.

Adding a listener for the "on loading finished" Blockly event helps by
resetting the trashcan's history when the event triggers.