Added shadow blocks for procedures
This commits adds shadow blocks to all blocks that are in the
functions category (or procedures if you prefer that name).

Additionally, a new block was added. It's a "return" block for functions.
If-return logic makes way more sense with an "if" block from the logic
category and a "return" block from the functions put inside it, as
opposed to the current solution which is to ignore the "if" block
altogether and provide an "if-return" block, which is what Blockly does.