Fix centering on block when showing search results
Google's team had a bug - the centerOnBlock function in WorkspaceSvg
was reusing a function for calculating a block's height and width.

However, that function was recursively adding the height of all
connected blocks. This is needed for the correct placement of blocks
when connecting and rendering them. It fails, however, when trying to
move the camera so it's on top of the block.

The fix is simple - add an optional parameter, which forces the function
to skip the recursive calculation of any connected blocks. The parameter
is only used by centerOnBlock and by nothing else, unless we find future
bugs.