Migrate core/utils.js to goog.module
Note that with core/utils.js in this form test/playground.html will
load and run correctly, but npm run build:compressed fails with many
errors like:

ERROR - [JSC_REFERENCE_TO_MODULE_GLOBAL_NAME] References to the global
name of a module are not allowed. Perhaps you meant exports?
  69|   const xy = new Blockly.utils.Coordinate(0, 0);
                               ^^^^^

I expect that this will be fixed in the next commit, when named
requires are used instead.