fix(blocks): Separate blocks namespace and dictionary objects
Use Blockly.blocks.* for blocks modules, leaving the Blockly.Blocks
name for the block dictionary object.

This solves a problem with advanced compilation, where (in the compiled
version) Blockly.Blocks is overwritten, replacing the dictionary
object defined in core/blocks.js with an empty namespace object created
by the provides of Blockly.Blocks.* in blocks/*.js.  This results in
some block definitions being created in the dictionary and some on the
namespace object, with chaos ensuing.