chore(deps): Remove deprecated babel-lint package
When dependabot submitted PR #5791, it became apparent that eslint
version >= 8 is incompatible with babel-eslint v10.1.0 despite the
latter not having a version restriction against it.  This is probably
because v10.1.0 is the last version of the package before it was
deprecated and renamed @babel/eslint-parser.

A little more investigation determined that there is no longer any
need to have eslint use a non-default parser for tests/mocha/**/*.js,
so the most straight forward solution is just to remove the dependency
entirely.