fix!(blocks): Have blocks_compressed.js export Blockly.blocks.all
Previously the value obtained by

    const blocks = require('blockly/blocks');  // Node.js

or

    import blocks from 'blockly/blocks.js';  // ES Modules

would be the block definitions dictionary (Blockly.Blocks).

Change this so that it is instead the export object from
Blockly.blocks.all.

This means you can now access loopTypes via:

    import blocks from 'blockly/blocks.js';
    blocks.loops.loopTypes.add('my_loop_blocktype');

This is a breaking change for any code which depended on the value
that was exported by blocks_compressed.js.

N.B.: at least in node.js, "import blocks from 'blockly/blocks'"
fails with the following error:

    Error [ERR_UNSUPPORTED_DIR_IMPORT]: Directory import '<...>/blocks'
    is not supported resolving ES modules imported from <...>
    Did you mean to import blockly/dist/blocks.js?

(The "dist" in the error message seems to be spurious.)