refactor!(blocks): Rename Blockly.blocks.all module to Blockly.blocks
This is a breaking change (only) because the exports object from the
`blocks_compressed.js` chunk will be accessed as `Blockly.blocks`
instead of `Blockly.blocks.all` when the chunk is loaded in a
browser via a `<script>` tag.  There will be no changes visible
when the chunk is loaded via ES module `import` or CJS `require`.