fix!(xml): Make appendDomToWorkspace take only WorkspaceSvg (#6162)
Turns out if one runs tsc on files in core/, the output will
cause Closure Compiler to detect type errors that were not detected
previously, despite apparently only whitespace changes between tsc's
input and output.

In this case, it detects that you can't call .getBlocksBoundingBox
on a Workspace object.

appendDomToWorkspace rearranges blocks on the workspace and is
only relevant to rendered workspaces. If you use this function in
headless mode, switch to using domToWorkspace instead.

BREAKING CHANGE: have appendDomToWorkspace take a WorkspaceSvg instead
of a Workspace as its second parameter.